### 2018-08-21 #907 @whedon
[PRE REVIEW]: `datastructures`: An R package for organisation and storage of data
**Submitting author:** @dirmeier (<a href="http://orcid.org/0000-0001-9795-3550">Simon Dirmeier</a>)
**Repository:** <a href="https://github.com/dirmeier/datastructures" target ="_blank">https://github.com/dirmeier/datastructures</a>
**Version:** v0.2.7
