### 2018-08-21 #907 @dirmeier
@karthik Hello, there are no conflicts of interest. I've found him through the JOSS reviewer list and thought his experience with R/C++ suits this project well.
