### 2018-08-22 dirmeier/datastructures#8 @fabian-s
utils::head, utils::stack, base::get, base::remove get overwritten without replacement
`datastructures` overwrites these fairly important functions/generics from always loaded packaged without replacement which will break a lot of other peoples' code:

Expected behavior:
```
> x <- 1:10
> get("x")
 [1]  1  2  3  4  5  6  7  8  9 10
> head(x, -1)
[1] 1 2 3 4 5 6 7 8 9
> stack(as.data.frame(x))
   values ind
1       1   x
2       2   x
3       3   x
4       4   x
5       5   x
6       6   x
7       7   x
8       8   x
9       9   x
10     10   x
> remove(x)
> x
Error: object 'x' not found
```
After `datastructures` is loaded:
```r
> library(datastructures)
> x <- 1:10
> get("x")
  Error in (function (classes, fdef, mtable)  :
  unable to find an inherited method for function ‘get’ for signature ‘"character", "missing", "missing"’
> head(x, -1)
Error in head(x, -1) : unused argument (-1)
> stack(as.data.frame(x))
Error in stack(as.data.frame(x)) : unused argument (as.data.frame(x))
> remove(x)
 Error in (function (classes, fdef, mtable)  :
  unable to find an inherited method for function ‘remove’ for signature ‘"integer", "missing", "missing"’
```

openjournals/joss-reviews#907
