### 2018-08-22 dirmeier/datastructures#8 @dirmeier
Got it. This can be solved with `remove.default` I guess?
