### 2018-08-22 dirmeier/datastructures#9 @fabian-s
`dequer` seems like relevant prior work in a very similar direction
IIUC, JOSS would like to see you mention relevant prior work that is similar,
so please mention/compare https://github.com/wrathematics/dequer

openjournals/joss-reviews#907
