### 2018-08-22 dirmeier/datastructures#10 @fabian-s
no examples/use cases
The documentation is OK'ish, but there is not  a single example showing inputs/outputs/use-cases on any of the help pages.

openjournals/joss-reviews#907
