### 2018-08-22 9b6b8e8
Renamed some functions and made them conform to base R
