+++ joss/paper.md
@@ -17,7 +17,7 @@ bibliography: paper.bib
 
 # Summary
 
-In computer science, data structures are objects for organisation and storage of data. They are often a key for efficient computation. So far the statistical computing language R [@rlang] only supports simple data structures such as lists, matrices and data tables, whereas other data structures such as priority queues, stacks or hashmaps are not supported. These data structures are essential in many computer science and statistics problems, like graph algorithms, string analysis or scheduling. In order to access a richer repertoire of data types, and thus enabling efficient computation of various problems, we propose the R package [datastructures](https://github.com/dirmeier/datastructures). The package uses Boost and STL data types in a C++ backend and exports these to R using the Rcpp [@rcpp] module system. 
+In computer science, data structures are objects for organisation and storage of data. They are often a key for efficient computation. So far the statistical computing language R [@rlang] only supports simple data structures such as lists, matrices and data tables, whereas other data structures such as priority queues, stacks or hashmaps are not supported. These data structures are essential in many computer science and statistics problems, like graph algorithms, string analysis or scheduling. Single packages, for example @hashmappackage or @dequerpackage, have already been proposed, although none of them support a wider array of data structures with a unified interface. In order to access a richer repertoire of data types, and thus enabling efficient computation of various problems, we propose the R package [datastructures](https://github.com/dirmeier/datastructures). The package uses Boost and STL data types in a C++ backend and exports these to R using the Rcpp [@rcpp] module system. 
 
 So far `datastructures` supports three different groups of data types. **Heaps** Fibonacci and binomial heaps can be used for priority queue operations, such as in (a version of) Dijkstra's algorithm for finding shortest paths. **Maps** Hash-, bi- and multimaps are associative arrays that establish key-value relationships. 
 **Deques** Stacks and queues are linked lists supporting the LIFO or FIFO principle, respectively.
