+++ joss/paper.bib
@@ -18,3 +18,17 @@
     url = {http://www.jstatsoft.org/v40/i08/},
     doi = {10.18637/jss.v040.i08},
   }
+
+@Manual{dequerpackage,
+    title = {{dequer}: Stacks, Queues, and Deques for R},
+    author = {Drew Schmidt},
+    note = {{R} package version 2.0-1},
+    url = {https://cran.r-project.org/package=dequer},
+  }
+
+@Manual{hashmappackage,
+    title = {hashmap: The Faster Hash Map},
+    author = {Nathan Russell},
+    note = {R package version 0.2.2},
+    url = {https://CRAN.R-project.org/package=hashmap},
+  }
