### 2018-08-22 6dbb030
Updates documentation
