+++ R/methods_values.R
@@ -20,7 +20,8 @@
 
 #' @title Get values from an object
 #'
-#' @description Extracts the values from a \code{map} object.
+#' @description Extracts the values from a data structure such as a \code{map}
+#'  or \code{heap} object.
 #'
 #' @export
 #' @docType methods
@@ -28,7 +29,26 @@
 #'
 #' @param obj  object to extract values from
 #'
-#' @return  returns the extracted values as vector
+#' @return  returns the extracted values as a \code{list} or, when primitive, as
+#'  a \code{vector}. In case of a \code{heap} also returns \code{key} and \code{handle}
+#'  of the heap node.
+#'
+#' @examples
+#'
+#'  # shows the values of a hashmap
+#'  h_map <- hashmap("integer")
+#'  h_map <- insert(h_map, seq(2), list(data.frame(a=1), 3))
+#'  values(h_map)
+#'
+#'  # shows the values of a multimap
+#'  m_map <- multimap("integer")
+#'  m_map[seq(2)] <- list(diag(2), rnorm(3))
+#'  values(m_map)
+#'
+#'  # shows the values of a heap
+#'  f_heap <- fibonacci_heap("integer")
+#'  f_heap <- insert(f_heap, 1:2, list(diag(2), rnorm(3)))
+#'  values(f_heap)
 #'
 setGeneric(
     "values",
