+++ R/methods_keys.R
@@ -30,6 +30,18 @@
 #'
 #' @return  returns the extracted keys as vector
 #'
+#' @examples
+#'
+#'  # returns the keys of a hashmap
+#'  h_map <- hashmap("numeric")
+#'  h_map[rnorm(3)] <- list(1, 2, 3)
+#'  keys(h_map)
+#'
+#'  # returns the keys of a multimap
+#'  m_map <- multimap("numeric")
+#'  m_map[c(1, 2, 1)] <- list(rnorm(1), rgamma(1, 1), rexp(1))
+#'  keys(m_map)
+#'
 setGeneric(
   "keys",
   function(obj) {
