+++ R/methods_erase.R
@@ -20,7 +20,7 @@
 
 #' @title Erase an entry from a map
 #'
-#' @description Erase a vector of key-value pair from a map object.
+#' @description Erase a vector of key-value pair from a \code{map} object.
 #'
 #' @export
 #' @docType methods
@@ -29,9 +29,23 @@
 #' @param obj  the object to pop an element from
 #' @param key  a vector of keys that should be removed
 #' @param value optionally a list of values needs to be supplied for some data
-#'  structures such as multimaps if entries should be removed.
+#'  structures such as \code{multimap}s if a single key-value pair should removed. If
+#'  not provided removes all key-value pairs with a specific key.
 
 #' @return  returns \code{obj} with removed values
+#' @examples
+#'
+#'  # erases keys from a \code{hashmap} or \code{bimap}
+#'  h_map <- hashmap()
+#'  h_map[letters] <- rnorm(length(letters))
+#'  h_map <- erase(h_map, "a")
+#'  h_map <- erase(h_map, letters[2:5])
+#'
+#'  # erases keys from a multimap
+#'  m_map <- multimap()
+#'  m_map[c("a", "a", "a", "b", "b", "c")] <- rep(1:2, 3)
+#'  m_map <- erase(m_map, "a")
+#'  m_map <- erase(m_map, "b", 1)
 #'
 setGeneric(
     "erase",

[…]
