+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: datastructures
 Type: Package
 Title: Implementation of Core Data Structures
-Version: 0.2.7
+Version: 0.2.8
 Authors@R: person("Simon", "Dirmeier",
                   email = "simon.dirmeier@web.de",
                   role = c("aut", "cre"))
