### 2018-08-22 dirmeier/datastructures#10 @dirmeier
@fabian-s Thanks for the feedback. I've updated examples for every method  and use cases and the landing page.
