+++ joss/paper.bib
@@ -22,6 +22,7 @@
 @Manual{dequerpackage,
     title = {{dequer}: Stacks, Queues, and Deques for R},
     author = {Drew Schmidt},
+    year = {2017},
     note = {{R} package version 2.0-1},
     url = {https://cran.r-project.org/package=dequer},
   }
@@ -29,6 +30,7 @@
 @Manual{hashmappackage,
     title = {hashmap: The Faster Hash Map},
     author = {Nathan Russell},
+    year = {2017},
     note = {R package version 0.2.2},
     url = {https://CRAN.R-project.org/package=hashmap},
   }
