### 2018-08-22 dirmeier/datastructures#9 @dirmeier
Added `dequer` and `hashmap` as references to the paper. DO you know whether there's a way to rebuild the document?
