### 2018-08-22 45d7fe0
Removes head instead of peek for maps
