+++ R/ds_deque_queue.R
@@ -32,6 +32,8 @@
 #'
 #' @slot .deque  \code{C++} object representing a deque
 #'
+#' @seealso  \code{\link{queue}} for creating a new \code{queue} object.
+#'
 setClass("queue", contains = "deque")
 
 
@@ -46,9 +48,9 @@ setClass("queue", contains = "deque")
 #' @return returns a new \code{queue} object
 #'
 #' @examples
-#'
 #'  # returns a new queue<SEXP>
 #'  q <- queue()
+#'
 queue <- function()
 {
     queue <- methods::new(queue_sexp)
