+++ R/ds_heap_fibonacci.R
@@ -34,6 +34,8 @@
 #' @slot .heap  \code{C++} object representing a heap
 #' @slot .key.class  the class of the keys
 #'
+#' @seealso  \code{\link{fibonacci_heap}} for creating a new \code{fibonacci_heap} object
+#'
 setClass("fibonacci_heap", contains = "heap")
 
 
@@ -50,7 +52,6 @@ setClass("fibonacci_heap", contains = "heap")
 #' @return returns a new \code{fibonacci_heap} object
 #'
 #' @examples
-#'
 #'  # creates a fibonacci_heap<character, SEXP>
 #'  f_heap <- fibonacci_heap()
 #'
