+++ R/methods_peek.R
@@ -44,6 +44,16 @@
 #'  b_heap <- insert(b_heap, letters[seq(3)], list(1, diag(3), rnorm(2)))
 #'  peek(b_heap)
 #'
+#'  # peeks into a \code{hashmap}
+#'  h_map <- hashmap()
+#'  h_map[letters] <- rnorm(length(letters))
+#'  peek(h_map)
+#'
+#'  # peeks into a \code{bimap}
+#'  b_map <- bimap("integer", "integer")
+#'  b_map[seq(10)] <- seq(10, 1)
+#'  peek(b_map)
+#'
 setGeneric(
     "peek",
     function(obj)
