### 2018-08-22 dirmeier/datastructures#8 @dirmeier
I got rid of `head`, `get` and `remove`. `stack` should work now, too. Thanks for all the feedback.
