### 2018-08-27 dirmeier/datastructures#9 @fabian-s
The way I read the guidelines, the bot is looking for a `paper.md` file in your repo (but maybe not in a subdirectory of your repo...?) I just tried to get their bot to re-generate the paper, let's see...
