### 2018-08-27 #910 @fabian-s
@dirmeier "generate pdf" might be an author command --
can you try to get whedon to re-generate your updated paper so I can check the missing checkmarks ;)
