### 2018-08-27 #910 @fabian-s
well damn! that worked!

@karthik: this is done (and very well done, too, @dirmeier!) except the fact that the version is off by one (says 0.2.7 in my checklist above, is now at 0.2.8) -- does that matter? can you correct from your side?
