### 2018-08-27 #910 @karthik
@fabian-s The Whedon commands are only for the editors. Thanks for checking in regarding the version issue. That is not a problem and generally to be expected. Authors will submit a version, but in the process of revisions end up bumping to something higher. That is typically what we expect.
